Improved GUIs + Bug Fixes
Added Information about other estates, including region information.

Made GUIs easier to read with new color guide. The working color guide is white for names, blue for information, green for true boolean, red for false boolean or for warning, gray for fields, white for integers, and gold for headings. Creating methods to automatically use these colors, and being able to edit them would be useful.

Fixed bug where subtracting costs only subtracts the first item. The subtractCosts method in DevelopmentManager does not return if it was successful or not.

Need to add check to see if the estate has enough food for all the developments before subtracting.