Added in the ability to Kick players from a nobility group.
Had to reset certain things to OfflinePlayer in order for compatibility purposes, we might have to reset most if not all of these in the future to check OfflinePlayer instead of player, as that could be bad, but hey we will figure it out.

Anyway seems to work after testing for ages.